fixed feature tabs pane
